Refactoring: Definition of task spawner
