+++ executorlib/task_scheduler/file/shared.py
@@ -7,7 +7,7 @@ from typing import Any, Callable, Optional
 from executorlib.standalone.command import get_cache_execute_command
 from executorlib.standalone.hdf import get_cache_files, get_output
 from executorlib.standalone.serialize import serialize_funct
-from executorlib.task_scheduler.file.subprocess_spawner import terminate_subprocess
+from executorlib.task_scheduler.file.spawner_subprocess import terminate_subprocess
 
 
 class FutureItem:
