+++ tests/test_cache_fileexecutor_mpi.py
@@ -5,7 +5,7 @@ import unittest
 
 try:
     from executorlib.task_scheduler.file.task_scheduler import FileTaskScheduler
-    from executorlib.task_scheduler.file.subprocess_spawner import execute_in_subprocess
+    from executorlib.task_scheduler.file.spawner_subprocess import execute_in_subprocess
 
     skip_h5py_test = False
 except ImportError:
