+++ tests/test_fluxclusterexecutor.py
@@ -12,7 +12,7 @@ try:
     import flux.job
     from executorlib import terminate_tasks_in_cache
     from executorlib.standalone.hdf import dump
-    from executorlib.task_scheduler.file.queue_spawner import execute_with_pysqa
+    from executorlib.task_scheduler.file.spawner_pysqa import execute_with_pysqa
     from executorlib.standalone.scheduler import terminate_with_pysqa
 
     skip_flux_test = "FLUX_URI" not in os.environ
